% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_venn_data.R
\name{process_venn_data}
\alias{process_venn_data}
\title{Process Venn Data}
\usage{
process_venn_data(filepath)
}
\arguments{
\item{filepath}{Path to the data file}
}
\value{
A tibble with processed Venn data
}
\description{
Process Venn Data
}
